Split Android graphics context implementations into their own files.
