+++ android/jni/AndroidEGLContext.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include "AndroidGraphicsContext.h"
+#include "Common/GL/GLInterfaceBase.h"
+
+class AndroidEGLGraphicsContext : public AndroidGraphicsContext {
+public:
+	AndroidEGLGraphicsContext() : draw_(nullptr), wnd_(nullptr), gl(nullptr) {}
+	bool Init(ANativeWindow *wnd, int desiredBackbufferSizeX, int desiredBackbufferSizeY, int backbufferFormat, int androidVersion) override;
+	void Shutdown() override;
+	void SwapBuffers() override;
+	void SwapInterval(int interval) override {}
+	void Resize() override {}
+	Draw::DrawContext *GetDrawContext() override {
+		return draw_;
+	}
+
+private:
+	Draw::DrawContext *draw_;
+	ANativeWindow *wnd_;
+	cInterfaceBase *gl;
+};
